+++ packages/steiger-plugin-fsd/src/_lib/collect-related-ts-configs.ts
@@ -1,5 +1,5 @@
 import { TSConfckParseResult } from 'tsconfck'
-import { dirname, posix } from 'node:path'
+import { dirname } from 'node:path'
 
 export type CollectRelatedTsConfigsPayload = {
   tsconfig: TSConfckParseResult['tsconfig']
@@ -106,16 +106,21 @@ function makeRelativePathAliasesAbsolute(
   firstConfigWithPaths: CollectRelatedTsConfigsPayload,
 ) {
   const { tsconfig: mergedConfig } = finalConfig
+  const {
+    compilerOptions: { paths, baseUrl },
+  } = mergedConfig
   const absolutePaths: Record<string, Array<string>> = {}
 
   if (!firstConfigWithPaths.tsconfigFile) {
     return mergedConfig.compilerOptions.paths
   }
 
-  for (const entries of Object.entries(mergedConfig.compilerOptions.paths)) {
+  for (const entries of Object.entries(paths)) {
     const [key, paths] = entries as [key: string, paths: Array<string>]
     absolutePaths[key] = paths.map((relativePath: string) =>
-      posix.resolve(dirname(firstConfigWithPaths.tsconfigFile!), relativePath),
+      baseUrl
+        ? resolve(dirname(firstConfigWithPaths.tsconfigFile!), baseUrl, relativePath)
+        : resolve(dirname(firstConfigWithPaths.tsconfigFile!), relativePath),
     )
   }
 
@@ -231,6 +236,88 @@ if (import.meta.vitest) {
           paths: {
             '~': ['/user/projects/project-0/src'],
             '~/*': ['/user/projects/project-0/src/*'],
+          },
+          strict: true,
+          noUncheckedIndexedAccess: false,
+          forceConsistentCasingInFileNames: true,
+          noImplicitOverride: true,
+          module: 'ESNext',
+          noEmit: true,
+        },
+      },
+    ]
+
+    expect(collectRelatedTsConfigs(payload)).toEqual(expectedResult)
+  })
+
+  test('correctly resolves paths if baseUrl is set', () => {
+    const payload: CollectRelatedTsConfigsPayload = {
+      tsconfigFile: resolve(joinFromRoot('user', 'projects', 'project-0', 'tsconfig.json')),
+      tsconfig: {
+        compilerOptions: {
+          baseUrl: './src',
+          paths: {
+            '~': ['./'],
+            '~/*': ['./*'],
+          },
+          strict: true,
+          noUncheckedIndexedAccess: false,
+          forceConsistentCasingInFileNames: true,
+          noImplicitOverride: true,
+          module: 'ESNext',
+          noEmit: true,
+        },
+      },
+    }
+
+    const expectedResult = [
+      {
+        compilerOptions: {
+          baseUrl: './src',
+          paths: {
+            '~': [resolve(joinFromRoot('user', 'projects', 'project-0', 'src'))],
+            '~/*': [resolve(joinFromRoot('user', 'projects', 'project-0', 'src', '*'))],
+          },
+          strict: true,
+          noUncheckedIndexedAccess: false,
+          forceConsistentCasingInFileNames: true,
+          noImplicitOverride: true,
+          module: 'ESNext',
+          noEmit: true,
+        },
+      },
+    ]
+
+    expect(collectRelatedTsConfigs(payload)).toEqual(expectedResult)
+  })
+
+  test('correctly resolves paths if baseUrl is set', () => {
+    const payload: CollectRelatedTsConfigsPayload = {
+      tsconfigFile: resolve(joinFromRoot('user', 'projects', 'project-0', 'tsconfig.json')),
+      tsconfig: {
+        compilerOptions: {
+          baseUrl: './src',
+          paths: {
+            '~': ['./'],
+            '~/*': ['./*'],
+          },
+          strict: true,
+          noUncheckedIndexedAccess: false,
+          forceConsistentCasingInFileNames: true,
+          noImplicitOverride: true,
+          module: 'ESNext',
+          noEmit: true,
+        },
+      },
+    }
+
+    const expectedResult = [
+      {
+        compilerOptions: {
+          baseUrl: './src',
+          paths: {
+            '~': [resolve(joinFromRoot('user', 'projects', 'project-0', 'src'))],
+            '~/*': [resolve(joinFromRoot('user', 'projects', 'project-0', 'src', '*'))],
           },
           strict: true,
           noUncheckedIndexedAccess: false,