+++ packages/steiger-plugin-fsd/src/index.ts
@@ -17,7 +17,7 @@ import segmentsByPurpose from './segments-by-purpose/index.js'
 import sharedLibGrouping from './shared-lib-grouping/index.js'
 import typoInLayerName from './typo-in-layer-name/index.js'
 import noProcesses from './no-processes/index.js'
-import packageJson from '../package.json'
+import packageJson from '../package.json' with { type: 'json' }
 
 const rules = [
   ambiguousSliceNames,
@@ -41,7 +41,7 @@ const rules = [
 
 const plugin = createPlugin({
   meta: {
-    name: 'steiger-plugin-fsd',
+    name: '@feature-sliced/steiger-plugin',
     version: packageJson.version,
   },
   ruleDefinitions: rules,